StrategyEditor Update

+++ 01_SW/Proccesing/StrategyEditor-main/CodePreviewWindow.pde
@@ -89,18 +89,20 @@ public class CodePreviewWindow extends PApplet {
 
     StrategyPoint first = StrategyEditor.points.get(0);
 
-    result += "Robot.setPosition("
-      + int(first.x_mm) + ", "
-      + int(first.y_mm) + ", "
-      + round(first.angleDeg) + ", Couleur_Team);\n";
+   result += "Robot.setPosition("
+    + int(first.x_mm) + ", "
+    + int(first.y_mm) + ", "
+    + int(round(first.angleDeg))
+    + ", Couleur_Team);\n";
 
     for (int i = 1; i < StrategyEditor.points.size(); i++) {
       StrategyPoint p = StrategyEditor.points.get(i);
 
-      result += "Robot.Robotgoto("
-        + int(p.x_mm) + ", "
-        + int(p.y_mm) + ", "
-        + round(p.angleDeg) + ", Couleur_Team, NORMALSPEED);";
+     result += "Robot.Robotgoto("
+  + int(p.x_mm) + ", "
+  + int(p.y_mm) + ", "
+  + int(round(p.angleDeg))
+  + ", Couleur_Team, NORMALSPEED);";
 
       if (p.poiName != null) {
         result += "  // Step " + i + " - POI " + p.poiName;
